Release: v2.19.0

+++ src/detection/gpu/gpu_driver_specific.h
@@ -41,7 +41,7 @@ typedef struct FFGpuDriverResult
     FFGPUMemory* memory;
     uint32_t* coreCount;
     FFGPUType* type;
-    double* frequency;
+    uint32_t* frequency;
 } FFGpuDriverResult;
 
 const char* ffDetectNvidiaGpuInfo(const FFGpuDriverCondition* cond, FFGpuDriverResult result, const char* soName);
